backend/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.9.1
asttokens==3.0.0
certifi==2025.8.3
charset-normalizer==3.4.3
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
decorator==5.2.1
distro==1.9.0
Django==5.2.6
django-extensions==4.1
djangorestframework==3.16.1
executing==2.2.1
fonttools==4.60.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
ipython==9.5.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jiter==0.11.0
joblib==1.5.2
kiwisolver==1.4.9
matplotlib==3.10.7
matplotlib-inline==0.1.7
numpy==2.3.4
openai==1.108.1
packaging==25.0
pandas==2.3.3
parso==0.8.5
pillow==11.3.0
prompt_toolkit==3.0.52
pure_eval==0.2.3
pydantic==2.11.9
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.5
scikit-learn==1.7.2
scipy==1.16.3
seaborn==0.13.2
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
stack-data==0.6.3
threadpoolctl==3.6.0
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.14
